XGBoost Tree-based Example Refactor and Refine  (#799)
* Refactor XGBoost Tree Learner

* Update README.md

minor typo corrections

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* Further refactors

* add one-shot training scheme

* reverse to the 5 sec fetch setting

* change config to 0 wait
