+++ examples/xgboost/job_configs/higgs_base/higgs_base/config/config_fed_server_bagging.json
@@ -35,7 +35,7 @@
         "min_clients": 5,
         "num_rounds": 101,
         "start_round": 0,
-        "wait_time_after_min_received": 10,
+        "wait_time_after_min_received": 0,
         "aggregator_id": "aggregator",
         "persistor_id": "persistor",
         "shareable_generator_id": "shareable_generator",
